refactor: reorganize predictor module directory structure
