+++ src/module/predictor/outpost/robot_state.cpp
@@ -1,4 +1,4 @@
-#include "outpost_robot_state.hpp"
+#include "robot_state.hpp"
 
 #include <algorithm>
 #include <array>
