+++ src/module/predictor/regular/robot_state.cpp
@@ -1,4 +1,4 @@
-#include "regular_robot_state.hpp"
+#include "robot_state.hpp"
 
 #include <algorithm>
 #include <cmath>
